+++ faraday/CHANGELOG.md
@@ -7,9 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First version.
 
-## [2] - 2022-04-19
+## 2 - 2022-04-19
 
-- Update extension to zap 2.11.1 and faraday v4
+- Update extension to zap 2.11.1 and faraday v4.
 
+## 3 - 2022-09-23
+
+- Update zap vulns ref to new format in faraday.
 
 [1]: https://github.com/infobyte/faraday_zap